refactor: mover enums a paquete separado y actualizar imports
